orders/src/routes/show.ts
import express, { Request, Response, NextFunction } from "express";
import {
  NotAuthorizedError,
  NotFoundError,
  requireAuth,
} from "@lxs_tickets/common";
import { Order } from "../models/order";

const router = express.Router();

router.get(
  "/api/orders/:orderId",
  requireAuth,
  async (req: Request, res: Response, next: NextFunction) => {
    const order = await Order.findById(req.params.orderId).populate("ticket");

    if (!order) {
      throw new NotFoundError();
    }

    if (order.userId !== req.currentUser!.id) {
      throw new NotAuthorizedError();
    }

    res.status(200).send(order);
  }
);

export { router as showsOrderRouter };
